requirements.txt
django==4.1.3
environs[django]==9.5.0
django-extensions==3.2.1
django-nested-admin==4.0.2